Migrate all raw  casts to infallible casts.rs wrappers (TODO #847)
9 files, 142+ cast sites across pillow-rs-freetype/src/.
All clippy::cast_* warnings eliminated (0 total vs 191 before).
Zero overhead: every wrapper is #[inline(always)] with debug_assert!

Files:
- casts.rs: fix usize_from_i64 debug_assert (was broken on 64-bit)
- fixed.rs, scaler.rs, font.rs, grays.rs: core math + render types
- tt/cmap.rs, tt/glyf.rs: table parser types
- autohint/loader.rs, autohint/latin.rs: autohinter point/edge types

Verification:
- cargo clippy: 0 cast_* warnings
- cargo test: 27686/27695 pass (9 unchanged autohinter failures)
- No behavioral change — pure mechanical replacement

+++ pillow-rs-freetype/src/casts.rs
@@ -64,8 +64,9 @@ pub(crate) fn u64_from_i64(x: i64) -> u64 {
 
 /// Infallible: i64 → usize. Values from division results or shifted coords.
 #[inline(always)]
+#[allow(clippy::cast_sign_loss)]
 pub(crate) fn usize_from_i64(x: i64) -> usize {
-    debug_assert!(x >= 0 && x <= usize::MAX as i64);
+    debug_assert!(x >= 0 && (x as u64) <= usize::MAX as u64);
     #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
     { x as usize }
 }
